Dependency Inject Amplifier
Now an amp isn't assumed - or required. Additionally, the task of
turning the amp on is not the responsibility of the guitar. The guitar
collaborates with the amp, but doesn't control it.

This also allows the guitar to work without an amplifier, which is a
reasonable use case. And trading out the type of amp can happen on
demand/instantiation, rather than requiring a code change. This allows
for a practice amp to have different output (nothing) than other amps,
which can be used for tests to silent and eliminate output.

The shuffle script is reworked to provide the guitar an amp. This
required changing the shapes from being arrays to being methods that
return arrays. The reason for this is that if we reuse the arrays, then
they are all picked at the time the script starts, which is before the
tube amp has time to warm up. As such, all the output shows that the amp
has no output. By instead invoking these methods fresh after the sleeps,
then the warm up time and volume is re-evaluated. It also is more
realistic, where while the shape may be replayed, it's still a fresh
invocation of the shape and not a reused copy of it.