Extract related behavior to separate class
This creates a Tuner, which is responsible for tuning a guitar. The
guitar still has a `tune` method on its public API, but the complexity
of doing the tuning is delegated to this other class.

This isn't an object that represents a real-life object, but still has
value being separate. Now the class it was extracted from, the guitar,
can focus on doing all of its guitar activities. Before it looked like
the main purpose of a guitar was to tune itself with the percentage of
code that was dedicated to tuning.