Sonic Pi Amp, Shuffle, and duration
This adds in a sonic pi amplifier that creates a ruby file that can be
passed to a sonic pi.

As part of this, we need notes to last for a certain period of time, and
as such, built in a duration.

This also changes the shuffle script to actually be a blues shuffle.
Before it was a blues progression, but didn't actually...shuffle. Which
is hard to know both when you can't hear it AND when notes don't have
any duration.