Shuffle class, separate scripts
This creates a class that knows how to complete a blues shuffle. It has
a play method that takes a block which is passed the results of each
measure. This way the different amps can process them differently as
needed.

This also adds in scripts for each of the amp types, rather than having
one generic script that was modified each time.